Move binding creation logic to `mesh_view_bindings.rs`
